+++ src/Console/Commands/Concerns/SavesFile.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace Aerni\Factory\Console\Commands\Concerns;
+
+use Illuminate\Support\Facades\File;
+use Illuminate\Support\Facades\Process;
+
+trait SavesFile
+{
+    protected function saveFile(string $path, string $contents): void
+    {
+        File::ensureDirectoryExists(dirname($path));
+
+        File::put($path, $contents);
+
+        Process::run('./vendor/bin/pint '.$path);
+    }
+}
